Head_pose/SC2/requirements.txt
absl-py==0.7.1
argcomplete==1.9.5
asn1crypto==0.24.0
astor==0.7.1
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
boto==2.48.0
cffi==1.12.3
colorama==0.4.1
crcmod==1.7
cryptography==2.6.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.5.0
entrypoints==0.3
fasteners==0.14.1
gast==0.2.2
gcs-oauth2-boto-plugin==2.3
google-apitools==0.5.27
google-reauth==0.1.0
grpcio==1.19.0
h5py==2.9.0
httplib2==0.12.3
imageio==2.5.0
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
mock==2.0.0
monotonic==1.5
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.3
notebook==5.7.5
numpy==1.16.2
oauth2client==4.1.3
opencv-contrib-python==4.1.0.25
pafy==0.5.4
pandocfilters==1.4.2
parso==0.3.4
pbr==5.1.3
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
Pygments==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.3.1
pyperclip==1.7.0
pyrsistent==0.14.11
python-chess==0.27.2
python-dateutil==2.8.0
python-gflags==3.1.2
pyu2f==0.1.4
PyWavelets==1.0.3
pywinpty==0.5.5
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.3
retry-decorator==1.1.0
rsa==4.0
scikit-image==0.15.0
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
SocksiPy-branch==1.1
tensorboard==1.12.1
tensorboardX==1.6
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
torch==1.1.0
torchvision==0.2.2.post3
tornado==6.0.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
youtube-dl==2019.4.30
